BUG: SequenceTests was not the correct place for test_annotate_matches_to
[CHANGED] all classes in file test_sequence are not annotatable. This should have
    triggered an error in testing but, mysteriously, it did not on Azure.
    I have moved this into the test_features module.
